+++ src/lib/repository/asessments.js
@@ -1,8 +1,11 @@
 import { PrismaClient } from "@prisma/client";
+import { verifyAccessToken } from "./middleware";
+import { AppError } from "../errors/AppError";
 
 const prisma = new PrismaClient();
 
-export async function createAsessment({ employee_id, priode_id, total_value, ranking }) {
+export async function createAsessment(token, { employee_id, priode_id, total_value, ranking }) {
+    verifyAccessToken(token);
     const newAsessment = await prisma.assessments.create({
         data: {
             employee_id,
@@ -27,7 +30,8 @@ export async function getAsessments() {
     });
 }
 
-export async function deleteAsessment(id) {
+export async function deleteAsessment(token, id) {
+    verifyAccessToken(token);
     return prisma.assessments.delete({
         where: {
             id
@@ -35,7 +39,8 @@ export async function deleteAsessment(id) {
     })
 }
 
-export async function updateAsessment(id, { employee_id, priode_id, total_value, ranking }) {
+export async function updateAsessment(token, id, { employee_id, priode_id, total_value, ranking }) {
+    verifyAccessToken(token);
     const newAsessment = await prisma.assessments.update({
         where: {
             id
